move gc class to circuits and clarify docstrings for fidelities
